llm_optimizer/requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.21.0
google-generativeai>=0.3.0

# Token counting
tiktoken>=0.5.0

# Utilities
numpy>=1.24.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# Optional: Redis for distributed caching
# redis>=5.0.0
